car_auctions_backend/requirements.txt
django==3.2.13
filters-django==1.0.5
django-bootstrap5==21.3
pillow
djangorestframework